Abstract events further into ComponentInteractEvent
