+++ core/src/main/java/discord4j/core/event/domain/interaction/ButtonInteractEvent.java
@@ -33,50 +33,9 @@ import reactor.core.publisher.Mono;
 import java.util.function.Consumer;
 
 @Experimental
-public class ButtonInteractEvent extends InteractionCreateEvent {
+public class ButtonInteractEvent extends ComponentInteractEvent {
 
     public ButtonInteractEvent(GatewayDiscordClient gateway, ShardInfo shardInfo, Interaction interaction) {
         super(gateway, shardInfo, interaction);
-    }
-
-    public String getCustomId() {
-        return getInteraction().getCommandInteraction() // yes, this is getCommandInteraction for buttons... thanks Discord
-                .flatMap(ApplicationCommandInteraction::getCustomId)
-                // note: custom_id is not guaranteed to present on buttons in general because of link buttons,
-                // but it is guaranteed to be present here, because we received an interaction for it
-                // (which doesn't happen for link buttons)
-                .orElseThrow(IllegalStateException::new);
-    }
-
-    // TODO: is this the right spec? needs rename
-    public Mono<FollowupHandler> edit(Consumer<? super InteractionApplicationCommandCallbackSpec> spec) {
-        return Mono.defer(
-                () -> {
-                    InteractionApplicationCommandCallbackSpec mutatedSpec =
-                            new InteractionApplicationCommandCallbackSpec();
-
-                    getClient().getRestClient().getRestResources()
-                            .getAllowedMentions()
-                            .ifPresent(mutatedSpec::setAllowedMentions);
-
-                    spec.accept(mutatedSpec);
-
-                    return respond(InteractionResponseType.UPDATE_MESSAGE, mutatedSpec.asRequest());
-                })
-                .thenReturn(followupHandler);
-    }
-
-    @Override
-    public Mono<FollowupHandler> deferResponse() {
-        return respond(InteractionResponseType.DEFERRED_UPDATE_MESSAGE, null);
-    }
-
-    @Override
-    public Mono<FollowupHandler> deferResponseEphemeral() {
-        InteractionApplicationCommandCallbackData data = InteractionApplicationCommandCallbackData.builder()
-                .flags(Message.Flag.EPHEMERAL.getFlag())
-                .build();
-
-        return respond(InteractionResponseType.DEFERRED_UPDATE_MESSAGE, data);
     }
 }
